refactor: static_cast除去・float64化・マジックナンバー定数化
- ateb_planner.cpp: float→double暗黙変換のためstatic_castを削除
- LocalPlannerConfig.msg: penalty_area_contractionをfloat32→float64に変更
  (他の距離定数と型を統一)
- robot_command_wrapper.hpp: setPenaltyAreaContractionの引数をdoubleに変更
- defender_session/total_defense_session: 0.5のマジックナンバーを
  kDefenderPenaltyAreaContraction定数に抽出しルール根拠をdocコメントに記載

+++ crane_local_planner/src/ateb_planner.cpp
@@ -163,8 +163,7 @@ auto ATEBPlanner::buildObstacles(uint8_t ego_id, const crane_msgs::msg::RobotCom
       needsExpandedPenaltyAreaOffset(world_model->getMsg().play_situation.command.value)
         ? PENALTY_AREA_OFFSET_STOP
         : PENALTY_AREA_OFFSET;
-    const double penalty_offset =
-      base_offset - static_cast<double>(cmd.local_planner_config.penalty_area_contraction);
+    const double penalty_offset = base_offset - cmd.local_planner_config.penalty_area_contraction;
     const auto our_pa = world_model->getOurPenaltyArea();
     const auto their_pa = world_model->getTheirPenaltyArea();
